requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# Segmentation models
ultralytics>=8.0.0
transformers>=4.35.0

# DINOv2 / Depth Anything dependencies
timm>=0.9.0
einops>=0.7.0

# Optional: for point cloud export
open3d>=0.17.0
tqdm>=4.65.0

# NOTE: zoedepth is embedded in ./zoedepth/ - no pip install needed!
